Move all Validator dependencies from Bank to blocktree_processor
